# contracts/client-agent-api.openapi.yaml
...
openapi: 3.1.0
info:
  title: Hermes DevOps Client Agent API
  version: 1.0.0
  description: |
    Runtime 对 Client Agent 的派单/取消/查询/设备/诊断接口。
    - 派单返回 202 仅表示"已入本地队列",执行进度经 callbacks-api 回传。
    - 幂等:重复 idempotency_key 返回已有任务状态,不重复执行。
    - 诊断只提供白名单探测,禁止任意 shell(红线 §14)。
security:
  - mtls: []
paths:
  /api/v1/tasks:
    post:
      operationId: dispatchTask
      summary: 派单(异步,仅确认入队)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskDispatchRequest'
      responses:
        '202':
          description: 已入本地队列(或重复幂等键时返回既有任务当前状态)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: 同 task_id 但幂等键不一致(契约冲突,拒绝)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 版本协商失败(Agent 版本低于 Runtime 要求的最低版本)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/tasks/{task_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getTask
      summary: 查询本地任务状态
      responses:
        '200':
          description: 任务当前状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: cancelTask
      summary: 取消任务(尽力而为:kill 设备进程 + 清理)
      responses:
        '202':
          description: 取消已受理,最终以 task-events/results 回调的终态为准
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/diagnostics:
    post:
      operationId: runDiagnostics
      summary: 白名单探测(非任意 shell)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiagnosticsRequest'
      responses:
        '200':
          description: 探测输出(截断到 max_bytes)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    mtls:
      type: mutualTLS
      description: Phase 3 强制双向 mTLS;Phase 1 允许自签部署,契约先行声明。
  responses:
    BadRequest:
      description: 请求不合法(未过 Schema 校验)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 任务不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    TaskDispatchRequest:
      type: object
      # 根级 additionalProperties 故意不设为 false(不同于内层 artifact/auth/
      # presigned_uploads 条目):契约"只加字段不删字段"演进,但 §14 严格
      # 校验会让每个新增 dispatch 字段(如 lease_id、upload_request_url)对
      # 已部署的旧 Agent 变成破坏性变更——旧 Agent 收到含新字段的载荷会直接
      # 400 schema_violation,而不是按设计忽略未知字段退化使用。放宽只从下一个
      # Agent 版本起生效,救不回已部署、仍是 additionalProperties:false 的旧
      # Agent;新 Runtime × 旧 Agent 组合仍需按发布顺序先升级 Agent
      # (见 docs/superpowers/specs/2026-07-29-on-demand-presign-design.md §7,
      # deploy/README.md 升级步骤)。
      required:
        - task_id
        - idempotency_key
        - attempt
        - artifact
        - manifest_digest
        - device_serial
        - callback_base_url
      properties:
        task_id: { type: string, minLength: 1, maxLength: 128 }
        idempotency_key:
          type: string
          minLength: 1
          maxLength: 256
          description: "{workflow_id}:{task_id}:{attempt},跨组件幂等的唯一依据"
        attempt: { type: integer, minimum: 1 }
        artifact:
          type: object
          additionalProperties: false
          required: [url, sha256, auth]
          properties:
            url: { type: string, format: uri, description: GitLab Generic Registry 确定性 URL }
            sha256: { type: string, pattern: '^[0-9a-f]{64}$', description: "整包校验和,下载后必须校验" }
            # CONTRACT-ISSUE: CLAUDE.md §8.1 原文为 auth: bearer|job_token,只有类型没有凭据。
            # 此处物化为 { type, token }:type 决定请求头形态(PRIVATE-TOKEN / JOB-TOKEN / HTTP Basic),token 为短期凭据。
            auth:
              type: object
              additionalProperties: false
              required: [type, token]
              properties:
                type: { enum: [bearer, job_token, basic] }
                token: { type: string, minLength: 1 }
                username:
                  type: string
                  description: |
                    basic(GitLab Deploy Token,原则 5 只读凭据)的 HTTP Basic
                    用户名;type=basic 时必填(运行时校验),其余类型忽略。
        manifest_digest:
          type: string
          pattern: '^[0-9a-f]{64}$'
          description: 包内 manifest.yaml 的 sha256,Client 解包后核对,防止包被替换
        device_serial: { type: string, minLength: 1 }
        callback_base_url: { type: string, format: uri }
        lease_id:
          type: string
          description: |
            租约所有权凭据(目标设计基线 v1.0 §10):Client 心跳续租时
            必须原样回传,失配即 LEASE_NOT_OWNED。只加字段不删字段,
            旧 Client 可暂不消费(过渡期心跳不续租)。
        lease_generation:
          type: integer
          minimum: 1
          description: 租约代数,与 lease_id 配对使用(同上)
        upload_request_url:
          type: string
          description: |
            按需签发端点的完整 URL(差距 #8),形如
            {callback_base_url}/callbacks/v1/upload-requests。
            Agent 收集完成后用它换取本次实际收集到的文件的预签名 PUT URL。
            为空 = Runtime 未启用按需签发,Agent 沿用 presigned_uploads[]。
        presigned_uploads:
          type: array
          default: []
          description: |
            MinIO 预签名上传地址,附件直传不经 Runtime(红线 §14)。
            新 Agent 仅在按需签发不可用(upload_request_url 为空或端点不可达)时用它兜底;
            本字段是回退路径的载体,不得移除。
          items:
            type: object
            additionalProperties: false
            required: [object_key, url]
            properties:
              object_key: { type: string, minLength: 1 }
              url: { type: string, format: uri }
              expires_at: { type: string, format: date-time }
    TaskStatus:
      type: object
      required: [task_id, state, attempt]
      properties:
        task_id: { type: string }
        idempotency_key: { type: string }
        attempt: { type: integer, minimum: 1 }
        state:
          description: Client 本地生命周期状态(§9 status 的 Client 可见子集)
          enum:
            - QUEUED
            - ACCEPTED
            - PREPARING
            - DOWNLOADING
            - DEPLOYING
            - RUNNING
            - COLLECTING
            - COMPLETED
            - FAILED
            - TIMEOUT
            - CANCELED
        detail: { type: string }
        updated_at: { type: string, format: date-time, description: UTC 毫秒精度 }
...
    DiagnosticsRequest:
      type: object
      additionalProperties: false
      required: [probe]
      properties:
        probe:
          description: 探测白名单,不可扩展为任意命令
          enum: [adb_devices, logcat_tail, df, getprop]
        args:
          type: object
          additionalProperties: false
          description: 受限参数,仅白名单键
          properties:
            serial: { type: string }
            lines: { type: integer, minimum: 1, maximum: 1000 }
            prop_name: { type: string, maxLength: 128 }
    DiagnosticsResponse:
      type: object
      required: [probe, output]
      properties:
        probe: { type: string }
        output: { type: string, description: 文本输出,超长截断 }
        truncated: { type: boolean, default: false }